workflows/requirements.txt
# ============================================
# Python 3.12 Compatible Requirements
# 
# ============================================

# ---- Core ML/DL Frameworks ----
torch>=2.2.0,<2.7.0
torchvision>=0.17.0,<0.22.0
torchaudio>=2.2.0,<2.7.0
pytorch-lightning>=2.2.0,<2.5.0
lightning-utilities>=0.11.0
torchmetrics>=1.3.0

# ---- Diffusion Models & Generative AI ----
diffusers>=0.27.0,<0.36.0
transformers>=4.40.0,<4.58.0
accelerate>=0.28.0,<1.12.0
peft>=0.10.0,<0.15.0
safetensors>=0.4.0
huggingface-hub>=0.22.0,<0.37.0

# ---- Image Processing ----
opencv-python-headless>=4.9.0,<4.12.0
pillow>=10.2.0,<11.0.0
scikit-image>=0.22.0,<0.25.0
imageio>=2.34.0,<2.38.0
imageio-ffmpeg>=0.4.9,<0.6.0
albumentations>=1.3.1,<1.5.0
imgaug>=0.4.0
kornia>=0.7.0,<0.8.0

# ---- Computer Vision & Segmentation ----
segment-anything>=1.0
open-clip-torch>=2.24.0,<2.30.0
timm>=0.9.16,<1.1.0
pycocotools>=2.0.7

# ---- Scientific Computing ----
numpy>=1.26.0,<2.0.0
scipy>=1.12.0,<1.15.0
pandas>=2.2.0,<2.3.0
scikit-learn>=1.4.0,<1.6.0
matplotlib>=3.8.0,<3.10.0
networkx>=3.2,<3.6

# ---- ONNX Deployment ----
onnx>=1.15.0,<1.18.0
onnxruntime>=1.17.0,<1.20.0
onnxconverter-common>=1.14.0

# ---- Configuration & Utilities ----
omegaconf>=2.3.0
hydra-core>=1.3.2,<1.4.0
pyyaml>=6.0.1
easydict>=1.10
einops>=0.7.0
tqdm>=4.66.0
joblib>=1.3.0
tabulate>=0.9.0

# ---- LLM/Agent Integration ----
langchain-mistralai>=0.1.0,<0.3.0
langchain-core>=0.2.0,<0.4.0
pydantic>=2.5.0,<2.10.0

# ---- Web & API ----
gradio>=4.20.0,<5.50.0
streamlit==1.32.0
streamlit-drawable-canvas==0.9.3
fastapi>=0.109.0,<0.122.0
uvicorn>=0.27.0,<0.39.0
requests>=2.31.0,<2.33.0
aiohttp>=3.9.0,<3.14.0
httpx>=0.27.0,<0.29.0
websockets>=12.0,<16.0
starlette>=0.36.0,<0.50.0
# streamlit-image-coordinates>=0.1.6

# ---- Visualization & Logging ----
tensorboard>=2.16.0,<2.18.0

# ---- Other Utilities ----
shapely>=2.0.0,<2.2.0
regex>=2023.12.0
ftfy>=6.1.0
webdataset>=0.2.60,<0.3.0
portalocker>=2.8.0
psutil>=5.9.0
memory-profiler>=0.61.0
protobuf>=4.25.0,<5.0.0
packaging>=23.2
wldhx.yadisk-direct>=0.0.6

# ---- Additional Dependencies ----
certifi>=2024.2.0
charset-normalizer>=3.3.0
colorama>=0.4.6
filelock>=3.13.0
fsspec>=2024.2.0
jinja2>=3.1.3
markupsafe>=2.1.5
mpmath>=1.3.0
networkx>=3.2
sympy>=1.12
typing-extensions>=4.10.0
urllib3>=2.2.0
click>=8.1.7
